requirements.txt
# FastAPI 관련
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic>=2.10.6

# 데이터베이스
pymysql==1.1.1

# Google Generative AI
google-genai>=1.52.0

# 의존성
anyio>=4.8.0

# 유틸리티
python-multipart==0.0.12
python-dotenv==1.0.1